hcproto: Improve hc_getgroups() to deal with getgroups() returning 0
The POSIX spec (2024 edition) says "The getgroups() function shall fill
in the array grouplist with the current supplementary group IDs of the
calling process. It is implementation-defined whether getgroups() also
returns the effective group ID in the grouplist array."  So in theory
getgroups() may return zero for a user that has no supplementary group.
However, the spec also says in the "CHANGE HISTORY" that "A return value
of 0 is not permitted, because {NGROUPS_MAX} cannot be 0. This is a FIPS
requirement."  So it becomes unclear whether getgroups() could possibly
return zero.

I have tested that getgroups() returns at least one (i.e., including the
primary group ID) on DragonFly, Linux, and Illumos (SmartOS).
Nevertheless, improve hc_getgroups() to explicitly deal with that
getgroups() may return zero.  Better safe than sorry.